first working version of the pgmax implementation of the rcn model. Trained on a small dataset of 20 images. Tested on 20 images.
